Resize Generate button slightly as hang workaround
Height of Generate button slightly increased as a temp workaround to current problems with the process hanging. The problem: "Interrupt" and "Skip" buttons are currently designed to be hidden until generate is pressed at which time they appear above the generate button, completely covering it. If the generation process hangs up, you're locked out of using the webui until you refresh the browser, thereby tediously losing all of your parameter settings. Until the issue with hangs is resolved or the ui is redesigned (hopefully the later since buttons covering buttons is bad practice in web interfaces) this fix causes the Generate button to peek out from below the Skip/Interrupt buttons so that it can continue to be used.